LayoutTests/http/tests/security/window-onerror-exception-in-iframe-expected.txt

CONSOLE MESSAGE: IframeException
Test that exceptions in iframe are not reported to the main frame window.onerror handler. window.onerror should print exactly one line.

Main frame window.onerror: MainFrameException at undefined:0
DONE
